requirements.txt
# 核心依赖：pip install -r requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.18
pymilvus>=2.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-docx>=0.8.11
pdfplumber>=0.10.3
jieba>=0.42.1
rank_bm25>=0.2.2
transformers>=4.35.2
torch>=2.1.0
requests>=2.31.0
python-multipart>=0.0.6
# 阿里云百炼依赖
dashscope>=1.14.0